Server Corrections
Made corrections in npm dependecies
