+++ .env
@@ -1,0 +1,1 @@
+GOOGLE_MAPS_API_KEY="bAhMUVLgqbtC1CuyEMJ0o2l-Y1LQ1Zq8M8sSOPEnRwg"
