+++ front_end/index.html
@@ -1,0 +1,207 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Google Map with Directions, Live Traffic, and Live Location</title>
+  <style>
+    /* Set the size of the map */
+    #map {
+      height: 70%;
+      width: 100%;
+    }
+
+    html, body {
+      height: 100%;
+      margin: 0;
+      display: flex;
+      flex-direction: column;
+    }
+
+    input, button {
+      width: 80%;
+      margin: 10px 0;
+      padding: 10px;
+      font-size: 16px;
+      box-sizing: border-box;
+    }
+
+    button {
+      background-color: #4CAF50;
+      color: white;
+      border: none;
+      cursor: pointer;
+    }
+
+    div {
+      padding: 10px;
+      text-align: center;
+      flex-shrink: 0;
+    }
+
+    h3 {
+      text-align: center;
+      font-size: 24px;
+      margin-bottom: 10px;
+    }
+    
+    @media (max-width: 600px) {
+      input, button {
+        width: 90%;
+      }
+      h3 {
+        font-size: 20px;
+      }
+    }
+
+    .navbar {
+      position: fixed;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 50px;
+      background-color: #d5e6cf;
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+      padding: 0 20px;
+      box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
+      z-index: 1000;
+    }
+
+    .menu-btn {
+      background: none;
+      border: none;
+      font-size: 20px;
+      cursor: pointer;
+      color: black;
+      position: relative;
+    }
+
+    .side-menu {
+      position: fixed;
+      top: 0;
+      left: -300px;
+      width: 250px;
+      height: 100%;
+      background-color: #81ce72;
+      box-shadow: 2px 0 5px rgba(0, 0, 0, 0.3);
+      overflow-y: auto;
+      transition: left 0.3s ease;
+      z-index: 9999;
+    }
+
+    .side-menu.open {
+      left: 0;
+    }
+
+    .side-menu ul {
+      list-style: none;
+      padding: 0;
+      margin: 20px 0;
+    }
+
+    .side-menu ul li {
+      padding: 10px 20px;
+    }
+
+    .side-menu ul li a {
+      text-decoration: none;
+      color: white;
+      font-size: 18px;
+    }
+
+    .side-menu ul li a:hover {
+      text-decoration: underline;
+    }
+  </style>
+</head>
+<body>
+  <div class="navbar">
+    <button class="menu-btn" id="menu-btn">&#9776;</button>
+    <span>Smart Shield</span>
+  </div>
+
+  <div class="side-menu" id="side-menu">
+    <input id="start" type="text" placeholder="Enter start location">
+    <input id="destination" type="text" placeholder="Enter destination">
+    <button onclick="calculateRoute()">Get Directions</button>
+  </div>
+  
+  <script>
+    const menuButton = document.getElementById('menu-btn');
+    const sideMenu = document.getElementById('side-menu');
+
+    menuButton.addEventListener('click', () => {
+      sideMenu.classList.toggle('open');
+    });
+  </script>
+
+  <div id="map"></div>
+
+  <script>
+    let map, directionsService, directionsRenderer, userLocationMarker;
+
+    function initMap() {
+      const center = { lat: 11.0168, lng: 76.9558 };  
+      map = new google.maps.Map(document.getElementById("map"), { zoom: 12, center: center });
+      directionsService = new google.maps.DirectionsService();
+      directionsRenderer = new google.maps.DirectionsRenderer({ map: map });
+
+      showUserLocation();
+
+      const urlParams = new URLSearchParams(window.location.search);
+      const start = urlParams.get('start');
+      const destination = urlParams.get('destination');
+
+      if (start && destination) {
+        document.getElementById("start").value = start;
+        document.getElementById("destination").value = destination;
+        calculateRoute();
+      }
+    }
+
+    function showUserLocation() {
+      if (navigator.geolocation) {
+        navigator.geolocation.getCurrentPosition(function(position) {
+          const userLatLng = { lat: position.coords.latitude, lng: position.coords.longitude };
+          map.setCenter(userLatLng);
+          map.setZoom(14);
+          userLocationMarker = new google.maps.Marker({ position: userLatLng, map: map, title: "You are here" });
+        });
+      } else {
+        alert("Geolocation service failed.");
+      }
+    }
+
+    function calculateRoute() {
+      const start = document.getElementById("start").value;
+      const destination = document.getElementById("destination").value;
+
+      if (!start || !destination) {
+        alert("Please enter both start and destination.");
+        return;
+      }
+
+      const request = {
+        origin: start,
+        destination: destination,
+        travelMode: google.maps.TravelMode.DRIVING
+      };
+
+      directionsService.route(request, function(result, status) {
+        if (status === google.maps.DirectionsStatus.OK) {
+          directionsRenderer.setDirections(result);
+        } else {
+          alert("Could not find route: " + status);
+        }
+      });
+    }
+  </script>
+
+  <script
+    src="https://maps.googleapis.com/maps/api/js?key=AIzaSyAue5c3sbdxrluKbDFYFQjIuncBrsZHRho&callback=initMap&libraries=places"
+    async
+    defer>
+  </script>
+
+</body>
+</html>
