--- html code/map.html
@@ -1,189 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Google Map with Directions</title>
-  <style>
-    /* Set the size of the map */
-    #map {
-      height: 70%; /* Reduced height to allow space for inputs on mobile */
-      width: 100%;
-    }
-
-    /* Set the height of the page to be full-screen */
-    html, body {
-      height: 100%;
-      margin: 0;
-      display: flex;
-      flex-direction: column;
-    }
-
-    input {
-      width: 80%;
-      margin: 10px 0;
-      padding: 10px;
-      font-size: 16px;
-      box-sizing: border-box;
-    }
-
-    button {
-      padding: 10px 20px;
-      font-size: 16px;
-      cursor: pointer;
-      width: 80%;
-      background-color: #4CAF50;
-      color: white;
-      border: none;
-    }
-
-    div {
-      padding: 10px;
-      text-align: center;
-      flex-shrink: 0;
-    }
-
-    h3 {
-      text-align: center;
-      font-size: 24px;
-      margin-bottom: 10px;
-    }
-
-    @media (max-width: 600px) {
-      input, button {
-        width: 90%; /* Make inputs and button take up more space on small screens */
-      }
-
-      h3 {
-        font-size: 20px; /* Reduce the heading size */
-      }
-
-      #map {
-        height: 60%; /* Reduce map height further on small screens */
-      }
-    }
-  </style>
-</head>
-<body>
-<h3>Google Map with Live Traffic and Directions</h3>
-
-<!-- Input fields for start and destination -->
-<div>
-  <input id="start" type="text" placeholder="Enter start location">
-  <input id="destination" type="text" placeholder="Enter destination">
-  <button onclick="calculateRoutes()">Get Directions</button>
-</div>
-
-<!-- The div element that will contain the map -->
-<div id="map"></div>
-
-<script>
-  let map, directionsService, directionsRendererShortest, directionsRendererLowestTraffic, trafficLayer;
-
-  // Initialize the map and services
-  function initMap() {
-    // Set the center to Coimbatore
-    const center = { lat: 11.0168, lng: 76.9558 };
-
-    // Create the map
-    map = new google.maps.Map(document.getElementById("map"), {
-      zoom: 12,
-      center: center,
-    });
-
-    // Directions service and renderers
-    directionsService = new google.maps.DirectionsService();
-
-    directionsRendererShortest = new google.maps.DirectionsRenderer({
-      map: map,
-      polylineOptions: { strokeColor: 'blue' },  // Blue line for shortest route
-    });
-
-    directionsRendererLowestTraffic = new google.maps.DirectionsRenderer({
-      map: map,
-      polylineOptions: { strokeColor: 'green' },  // Green line for low-traffic route
-    });
-
-    // Traffic Layer for live traffic data
-    trafficLayer = new google.maps.TrafficLayer();
-    trafficLayer.setMap(map);
-  }
-
-  // Function to calculate and display both the shortest route and the lowest traffic route
-  function calculateRoutes() {
-    const start = document.getElementById("start").value;
-    const destination = document.getElementById("destination").value;
-
-    // Ensure both start and destination are provided
-    if (!start || !destination) {
-      alert("Please enter both start and destination locations.");
-      return;
-    }
-
-    // Request for the shortest route (no traffic consideration)
-    const shortestRouteRequest = {
-      origin: start,
-      destination: destination,
-      travelMode: google.maps.TravelMode.DRIVING,
-    };
-
-    // Request for the lowest traffic route
-    const lowestTrafficRouteRequest = {
-      origin: start,
-      destination: destination,
-      travelMode: google.maps.TravelMode.DRIVING,
-      drivingOptions: {
-        departureTime: new Date(),  // Use current time for live traffic
-        trafficModel: google.maps.TrafficModel.BEST_GUESS, // "BEST_GUESS" accounts for traffic
-      },
-    };
-
-    // Calculate and display the shortest route
-    directionsService.route(shortestRouteRequest, function(result, status) {
-      if (status == google.maps.DirectionsStatus.OK) {
-        directionsRendererShortest.setDirections(result);
-        sendEmailWithLink(start, destination);  // Send email with Google Maps link after route is found
-      } else {
-        alert("Could not find shortest route: " + status);
-      }
-    });
-
-    // Calculate and display the lowest traffic route
-    directionsService.route(lowestTrafficRouteRequest, function(result, status) {
-      if (status == google.maps.DirectionsStatus.OK) {
-        directionsRendererLowestTraffic.setDirections(result);
-      } else {
-        alert("Could not find lowest traffic route: " + status);
-      }
-    });
-  }
-
-  // Function to send email with Google Maps link
-  function sendEmailWithLink(start, destination) {
-    const googleMapsLink = `https://www.google.com/maps/dir/?api=1&origin=${encodeURIComponent(start)}&destination=${encodeURIComponent(destination)}&travelmode=driving`;
-
-    fetch('http://localhost:3000/send-email', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        start: start,
-        destination: destination,
-        link: googleMapsLink,
-      }),
-    })
-    .then(response => response.json())
-    .then(data => alert(data.message))
-    .catch(error => console.error('Error:', error));
-  }
-</script>
-
-<!-- Load the Google Maps API script with your API key -->
-<script
-  src="https://maps.googleapis.com/maps/api/js?key=AIzaSyAue5c3sbdxrluKbDFYFQjIuncBrsZHRho&callback=initMap&libraries=places"
-  async
-  defer
-></script>
-</body>
-</html>
